+++ src/sub-clients/ageRange/client/AgeRangeClient.ts
@@ -1,0 +1,5 @@
+import { AgeRange } from '../model/AgeRange';
+
+export interface AgeRangeClient {
+  getAll(): Promise<AgeRange[]>;
+}