+++ src/sub-clients/collections/model/CollectionEntity.ts
@@ -1,4 +1,4 @@
-import { AgeRange } from '../../common/model/AgeRange';
+import { AgeRanges } from '../../common/model/AgeRanges';
 import { EntityWithLinks } from '../../common/model/common';
 import { LinkEntity } from '../../common/model/LinkEntity';
 
@@ -12,7 +12,7 @@ export interface CollectionEntity extends EntityWithLinks {
   mine: boolean;
   createdBy: string;
   subjects: any[];
-  ageRange: AgeRange | null;
+  ageRange: AgeRanges | null;
   description: string | null;
   attachments: AttachmentEntity[];
   _links: {