+++ src/sub-clients/contentPartners/resources/ContentPartnerResource.ts
@@ -1,11 +1,11 @@
-import { AgeRange } from '../../common/model/AgeRange';
+import { AgeRanges } from '../../common/model/AgeRanges';
 import { IngestDetailsResource } from './IngestDetailsResource';
 
 export interface ContentPartnerResource {
   id: string;
   name: string;
   official: boolean;
-  ageRange?: AgeRange;
+  ageRange?: AgeRanges;
   currency?: string;
   legalRestriction?: any;
   distributionMethods: any[];