+++ src/sub-clients/educationalAgeRanges/client/EduAgeRangesClient.ts
@@ -1,5 +1,0 @@
-import { EduAgeRange } from '../model/EduAgeRange';
-
-export interface EduAgeRangesClient {
-  getAll(): Promise<EduAgeRange[]>;
-}